Refactor CLI output contract and workflow subcommands
