+++ crates/pcl/cli/tests/verify_cli.rs
@@ -38,8 +38,11 @@ fn assert_verify_success(output: std::process::Output) {
     );
 
     let stdout = String::from_utf8(output.stdout).expect("utf-8 stdout");
-    let summary: serde_json::Value = serde_json::from_str(&stdout).expect("json summary");
-    assert_eq!(summary["status"], "success");
+    let envelope: serde_json::Value = serde_json::from_str(&stdout).expect("json envelope");
+    assert_eq!(envelope["status"], "ok");
+    assert_eq!(envelope["schema_version"], "pcl.envelope.v1");
+    let summary = &envelope["data"];
+    assert_eq!(summary["outcome"], "success");
     assert_eq!(summary["total"], 1);
     assert_eq!(summary["passed"], 1);
     assert_eq!(summary["failed"], 0);
@@ -113,8 +116,11 @@ fn apply_dry_run_builds_and_verifies_fixture_payload_without_api() {
 
     assert_command_success(&output, "pcl apply --dry-run");
     let stdout = String::from_utf8(output.stdout).expect("utf-8 stdout");
-    let summary: serde_json::Value = serde_json::from_str(&stdout).expect("json summary");
-    assert_eq!(summary["status"], "dry_run");
+    let envelope: serde_json::Value = serde_json::from_str(&stdout).expect("json envelope");
+    assert_eq!(envelope["status"], "ok");
+    assert_eq!(envelope["schema_version"], "pcl.envelope.v1");
+    let summary = &envelope["data"];
+    assert_eq!(summary["outcome"], "dry_run");
     assert_eq!(
         summary["project_id"],
         "550e8400-e29b-41d4-a716-446655440000"
